// BookStoreFrontend/src/app/Model/book/book.module.ts
import { NgModule } from '@angular/core';
import { CommonModule } from '@angular/common';



@NgModule({
  declarations: [],
  imports: [
    CommonModule
  ]
})
export class BookModule {
  [x: string]: any;

  bookId: number;
  BookModule: string;
  bookDetails: string;
  authorName: string;
  bookName: string;
  price: number;
  priceSale:number;
  noOfBooks: number;
  image: string;
  createdDateAndTime: Date;
  status: string;
 }